% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{chain_convergence}
\alias{chain_convergence}
\title{Perform checks for convergence of multiple Markov chains.}
\usage{
chain_convergence(chain)
}
\arguments{
\item{chain}{(array) of MCMC samples, \code{N} rows (sampled) by
\code{M} columns (variables).}
}
\value{
The values of R.hat for each of the \code{M} variables.
}
\description{
\code{chain_convergence} checks convergence of multiple Markov chains.
}
\details{
Uses Gelman & Rubin's R.hat for each parameter, and also a visual check of 
the 80\% regions for each parameter.
}
\seealso{
\code{\link{Rhat}}
}
