% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{contour_matrix}
\alias{contour_matrix}
\title{Produce a matrix of contours plots from a data array}
\usage{
contour_matrix(theta, ranges = NULL, cex = 1, prob.levels = c(0.9,
  0.95), labels = colnames(theta), upper = NA, dot.level = NULL,
  breaks = 30, smooth1d = TRUE, smooth2d = TRUE, npix = 100,
  cex.lab = 1, cex.axis = 1, prob1d = NULL, sigma = FALSE,
  jittr = FALSE, thin = NULL, pch = 1, col = "black",
  plot.image = FALSE, plot.1dlines = TRUE, fill.col = "steelblue3",
  ...)
}
\arguments{
\item{theta}{(array) \code{N} by \code{M} array of parameter values}

\item{ranges}{(array) \code{M} by \code{2} array of ranges for the plots}

\item{cex}{(float) character expansion factor.}

\item{prob.levels}{(float) probability levels, e.g. to plot contours enclosing
90\% and 95\% of the mass on 2D distributions}

\item{labels}{(array of strings) names of the variables}

\item{upper}{(string) plot also in upper-right triangle. See notes.}

\item{dot.level}{(integer) draw dots outside of which contour? (1,2,...)}

\item{breaks}{(integer) how many bins for 1D histograms}

\item{smooth1d}{(logical) plot smoothed distributions, rather than 
histograms on the leading diagonal?}

\item{smooth2d}{(logical) use smooth, kernel density estimates, rather 
than 2D histograms?}

\item{npix}{(integer) use \code{npix*npix} grid of pixels for 
computing 2D smoothed distributions (resolution of 
contour maps)}

\item{cex.lab}{(float) expansion factor for axis labels.}

\item{cex.axis}{(float) expansion factor for axis names.}

\item{prob1d}{(float array) probability levels at which to mark 
intervals on 1D distributions, e.g. c(0.683, 0.90).}

\item{sigma}{(logical) are probability levels \code{prob1d} and 
\code{prob.levels} given in terms of sigmas?
(If \code{FALSE} - the default - then these are 
probabilities.)}

\item{jittr}{(logical) add 'jitter' to points to reduce overlap.}

\item{thin}{(integer) a factor by which to 'thin out' data input data
i.e. plot only a fraction \code{1/thin} of the data 
points.}

\item{col}{(string) colour of the data points.}

\item{plot.image}{(logical) Plot a colour image underneath the contours?}

\item{plot.1dlines}{(logical) Mark intervals on 1D plots (histogram/density)?}

\item{fill.col}{(string) Colour to use under the histogram.}

\item{...}{(anything) any other graphical keywords to be passed to 
\code{plot(...)}}
}
\value{
Array of the 1D intervals for each parameter, or NULL.
}
\description{
\code{contour_matrix} generates an \code{M*M} matrix of scatter/density plots.
                 
Given an array of data with \code{M} columns (variables) and \code{N} rows 
(observations), produce a matrix of plots showing contours for each pair of 
parameters. The \code{pairs()} function shows a matrix of scatter plots for 
each pair of variables, here the scatter plots are replaced by contour plots.
The contours are produced by using \code{MASS:kde2d} to produce a 2D Gaussian
kernal density estimate, and finding the contours that encolse a fraction
(prob) of the total density.
}
\section{Notes}{

The input \code{chain} should be a list such as produced by 
  \code{gw_sampler} or \code{mh_sampler} that contains the following:
\describe{ 
\item{theta}{(array) \code{n * ndim} array of posterior samples
            \code{n} samples of \code{ndim} vectors of parameters}
\item{method}{(string) name of MCMC method used}
\item{nchains}{number of chains/walkers used}
}

The \code{upper} parameter determines what is to be shown in the upper-right
corner of the plot. The options are: \itemize{
                     \item{\code{"NA"} - no plot} 
                     \item{\code{"points"} - scatter plot of points}
                     \item{\code{"image"} - intensity map} 
                     \item{\code{"contour"} - contour map}
                     }
                     
You may notice that the fraction of points plotted outside the 
\code{dot.levels} contour line is often smaller than
\code{prob.levels[dot.levels]}. This is usually true for finite \code{N} and
is a natural consequence of plotting contours that enclose a fraction of the
smoothed density rather than a fraction of the points.
}

\examples{
n <- 10000
x <- rchisq(n, df=2)
y <- x + rnorm(n)
z <- rnorm(n, sd=1)
dat <- cbind(x,y,z)
contour_matrix(dat, smooth1d=TRUE)
contour_matrix(dat, prob.levels=c(1,2), cex=0.5, npix=100, sigma=TRUE,
          dot.level=2, prob1d=c(1, 2), smooth1d=TRUE, cex.lab=1.3,
          plot.1dlines = TRUE)

}
\seealso{
\code{\link{gw_sampler}}, \code{\link{mh_sampler}}, 
  \code{\link{get_levels2d}}, \code{\link{get_levels1d}}, 
  \code{\link{interp_image}}, \code{\link{hist2d}}, \code{\link{plot_density}},
  \code{\link{plot_density_contours}}
}
