% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{get_levels1d}
\alias{get_levels1d}
\title{Find lower, upper intervals enclosing certain masses of a histogram.}
\usage{
get_levels1d(x, prob = 0.95)
}
\arguments{
\item{x}{(vector) x data values.}

\item{prob}{(vector) probability masses at which to define levels.}
}
\value{
Vector of the x values (two for each element of \code{prob}). 
 If \code{prob = 0.95} then the lower, upper values enclose 95\% of the mass
 of x.
}
\description{
\code{get_levels1d} returns values bounding specified masses.
}
\details{
Function to compute the lower and upper limits of intevals containing a
fraction (\code{prob}) of the mass of the points \code{x}. If \code{x} are
input data points, we compute the lower and upper quantiles. For for
\code{prob = 0.95} we find the 0.025 and 0.975 quantiles, between which are 
95\% of the data points.
}
\examples{
n <- 1000
x <- rnorm(n)
get_levels1d(x, prob = c(0.5, 0.9))

}
\seealso{
\code{\link{mcmc.conv}}
}
