% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{get_levels2d}
\alias{get_levels2d}
\title{Compute contours encircling specified masses.}
\usage{
get_levels2d(img, prob = 0.95)
}
\arguments{
\item{img}{(list) containing three components: \code{x, y, z}}

\item{prob}{(vector) probability masses at which to define levels.}
}
\value{
Vector of the density levels (one for each element of \code{prob}). 
 If \code{prob = 0.95} then a contour a level contains 95\% of the mass.
}
\description{
\code{get_levels2d} returns density levels enclosing specified masses.
}
\details{
Function to compute the density levels at which to draw contour lines in
order to enclose a fraction \code{prob} of the mass.
}
\section{Notes}{

The input \code{img} should be a list with three components:
 \code{img$z} is a vector of \code{nx} evenly spaced \code{x} positions,
 \code{img$y} is a vector of \code{ny} evenly spaced \code{y} position,
 \code{img$z} is an \code{nx*ny} array of \code{z} values at the 
 \code{(x, y)} coordinates.
}

\examples{
n <- 1000
x <- rnorm(n)
y <- x + rnorm(n)
img <- hist2d(x, y, npix=11)
get_levels2d(img, prob = c(0.5, 0.9))

}
\seealso{
\code{\link{mcmc.conv}}
}
