% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{hist2d}
\alias{hist2d}
\title{Bin a set of (x,y) points into a 2D histogram}
\usage{
hist2d(x, y, npix = 25, lims = c(range(x), range(y)), prob = FALSE)
}
\arguments{
\item{npix}{number of bins on each axis (pixels = \code{npix^2})}

\item{lims}{Four-element vector for \code{x, y} limits of the histogram}

\item{prob}{(logical) make the area under the histogram equal to 1?}
}
\value{
List containing components \code{x, y, z}.
}
\description{
\code{hist2d} produce a 2D histogram from \code{x, y} points.
}
\details{
From an set of \code{(x, y)} positions, compute a 2D histogram.
Count the number of input points falling within each pixel on
and evenly spaced grid points of \code{pix.x = i*dx}, \code{pix.y = j*dy}. 
return this values as \code{z}.
}
\section{Notes}{

The output \code{img} is a list with three components:
 \code{z} is a vector of \code{nx} evenly spaced \code{x} positions,
 \code{y} is a vector of \code{ny} evenly spaced \code{y} position,
 \code{z} is an \code{nx*ny} array of \code{z} values at the 
 \code{(x, y)} coordinates.
}

\examples{
n <- 1000
x <- rnorm(n)
y <- x + rnorm(n)
img <- hist2d(x, y, npix=11)
image(img$x, img$y, log10(img$z))
points(x,y)

}
\seealso{
\code{\link{mcmc.conv}}
}
