% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mcmc.diag.plot}
\alias{mcmc.diag.plot}
\title{Plots for assessing MCMC output.}
\usage{
mcmc.diag.plot(chain, max.chain = 20)
}
\arguments{
\item{chain}{(list) containing output data from an MCMC, including}

\item{max.chains}{maximum no. chains to overlay on trace plot}
}
\value{
none
}
\description{
\code{mcmc.diag.plot} Generate plots to help assess the convergence 
                 of MCMC output.
                 
Generate plots of chain traces (variable values vs. iteration), 
auto-correlation plots and 1D densities (histograms).
}
\section{Further details}{

The input \code{chain} should be a list such as produced by 
  \code{gw.mcmc} or \code{mh.mcmc} that contains the following:
\describe{ 
\item{theta}{(array) n * ndim array of posterior samples
            n samples of ndim vectors of parameters}
\item{method}{(string) name of MCMC method used}
\item{nwalkers}{number of walkers used (if method = 'gw.mcmc')}
\item{nchains}{number of walkers used (if method = 'mh.mcmc')}
}
}

\examples{
my_posterior <- function(theta) {
  cov <- matrix(c(1,0.98,0.8,0.98,1.0,0.97,0.8,0.97,2.0), nrow = 3)
  logP <- mvtnorm::dmvnorm(theta, mean = c(-1, 2, 0), sigma = cov, log = TRUE)
  return(logP)
}
chain <- gw.mcmc(my_posterior, theta.0 = c(0,0,0), nsteps = 10e4, burn.in = 1e4) 
mcmc.diag.plot(chain)

}
\seealso{
\code{\link{gw.mcmc}}
}
