% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_density}
\alias{plot_density}
\title{Produce a 1D density/histogram plot given data values \code{x}.}
\usage{
plot_density(x, breaks = 30, xlim, prob = NULL, smooth = FALSE,
  fill.col = NULL, plot.lines = TRUE, xlab = "", ylab = "", main = "",
  ...)
}
\arguments{
\item{x}{(vector) x data values.}

\item{breaks}{(integer) how many bins for 1D histograms}

\item{xlim}{(vector) The limits of the x axis (x1, x2).}

\item{fill.col}{(string) Colour to use under the histogram.}

\item{xlab, ylab}{(strings) Labels for x, y axes.}

\item{...}{(anything) any other graphical keywords to be passed to 
\code{plot(...)}}
}
\value{
None.
}
\description{
\code{plot_density} generates a 1D density/histogram plot.
}
\details{
Produce a single histogram or density plot, for use with \code{contour_matrix}.
}
\examples{
n <- 1000
x <- rnorm(n)
y <- x + rnorm(n)
img <- hist2d(x, y, npix = 40)
plot_density(x, xlim = c(-4,4))
axis(1)

}
\seealso{
\code{\link{chain_convergence}}, \code{\link{contour_matrix}}
}
