% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_density_contours}
\alias{plot_density_contours}
\title{Produce a 2D density/histogram plot given data values \code{x, y}.}
\usage{
plot_density_contours(x, y, xlim, ylim, npix = 25,
  prob.levels = c(0.683, 0.9), plot.dots = TRUE,
  plot.contours = TRUE, plot.image = FALSE, dot.level = NULL,
  jittr = FALSE, col = "black", pch = 1, cex = 1,
  smooth2d = TRUE, xlab = "", ylab = "", ...)
}
\arguments{
\item{x, y}{(vectors) x and y data values (equal length vectors).}

\item{xlim, ylim}{(vectors) The (x, y) limits of the axes (x1, x2), (y1, y2).}

\item{npix}{(integer) use \code{npix*npix} grid of pixels for 
computing 2D smoothed distributions (resolution of 
contour maps)}

\item{prob.levels}{(float) probability levels, e.g. to plot contours enclosing
90\% and 95\% of the mass on 2D distributions}

\item{plot.dots}{(logical) Plot points outside of outer contour?}

\item{plot.contours}{(logical) Plot the contours?}

\item{plot.image}{(logical) Plot a colour image underneath the contours?}

\item{dot.level}{(integer) draw dots outside of which contour? (1,2,...)}

\item{jittr}{(logical) add 'jitter' to points to reduce overlap.}

\item{col}{(string) colour of the data points.}

\item{pch}{(integer) Plotting 'character', i.e. symbol to use. Same as 
'base' graphics.}

\item{cex}{(float) character expansion factor.}

\item{smooth2d}{(logical) use smooth, kernel density estimates, rather 
than 2D histograms?}

\item{xlab, ylab}{(strings) Labels for x, y axes.}

\item{...}{(anything) any other graphical keywords to be passed to 
\code{plot(...)}}
}
\value{
None.
}
\description{
\code{plot_density_contours} generates a 2D density/histogram plot.
}
\details{
Given vectors \code{x, y} specifying the coordinates of data points,
\code{plot_density_contours} generates a 2D density plot. What is plotted are contours
enclosing specified masses of the distribution, and points outside of the
outermost contour are marked individually. The contours are based on
either a smooth 2D kernel estimate or an a (rough) 2D histogram.
}
\examples{
n <- 1000
x <- rnorm(n)
y <- x + rnorm(n)
img <- hist2d(x, y, npix = 40)
plot_density_contours(x,y, xlim = c(-4, 4), ylim = c(-4, 4))
axis(1); axis(2)

}
\seealso{
\code{\link{chain_convergence}}, \code{\link{contour_matrix}}
}
