% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{Rhat}
\alias{Rhat}
\title{compute the Gelman & Rubin R.hat statistic.}
\usage{
Rhat(theta)
}
\arguments{
\item{theta}{(array) of samples of one parameters. \code{ncycles} rows by
\code{nchains} columns. We compare the within-chain and between-chain
variances.}
}
\value{
The R.hat statistic (scalar).
}
\description{
\code{Rhat} computes the Gelman & Rubin R.hat statistic.
Given an array of \code{theta} values, for some parameter, produced by
several chains, compute the R.hat statistic as a check for convergence. The
\code{R.hat} statistic (Gelman & Rubin 1992) should be close to zero if the
chains are converging. See also Gelman et al. (2004, sect 11.6)
}
\seealso{
\code{\link{mcmc.conv}}
}
