% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{chain_convergence}
\alias{chain_convergence}
\title{Perform checks for convergence of multiple Markov chains.}
\usage{
chain_convergence(chain)
}
\arguments{
\item{chain}{(array) of MCMC samples, \code{N} rows (sampled) by
\code{M} columns (variables).}
}
\value{
The values of R.hat for each of the \code{M} variables.
}
\description{
\code{chain_convergence} checks convergence of multiple Markov chains.
}
\details{
Return and plot Gelman & Rubin's R.hat statistic for each parameter, 
comparing within-chain and between-chain variances.
Plot a comparison between the 80\% regions for each parameter.
}
\section{Notes}{

This requires the input array be a two-dimensional array such as produced by
 mh_sampler or gw_sampler, with results from each chain/walker merged.
 The intervals are scaled so that each variable has mean 0 and std.dev 1
 to make it easier to compare variables which might have very different 
 scales.
 
 If the chains are `well mixed' R.hat should be close to 1.0 (ideally <1.1)
 and the intervals for each chain should share a lot of overlap. Note that
 this is more useful for the output of the MH method. (The GW method works
 best with a large ensemble of walkers - nwalkers >= 50 - but requires fewer
 iterations of the full ensemble, so the inter-walker comparisons are less
 useful.)
}

\seealso{
\code{\link{Rhat}}
}
